+++ CHANGELOG.md
@@ -25,3 +25,4 @@ Second version after feedback
 - Update: phone number format
 - Update: CTRL + S to save service form
 - Feature: scroll into view on form validation error & minor changes
+- Feature: order of columns in table of repairs